requirements.txt
Flask
requests
python-dotenv
openai
playwright
